key/requirements.txt
apscheduler
flask
flask_restful
PyYAML
pymysql
nltk
pandas
jinja2
